+++ documentation/website/.dir-template/src/docs/intro/change-log.md
@@ -25,6 +25,7 @@ Starred releases are most stable ones.
 + Add    NutsConstants.IdProperties.DESKTOP_ENVIRONMENT -> NutsConstants.IdProperties.PROPERTIES
 + Rename NutsRepositoryDB::getRepositoryNameByURL -> NutsRepositoryDB::getRepositoryNameByLocation
 + Rename NutsRepositoryDB::getRepositoryURLByName -> NutsRepositoryDB::getRepositoryLocationByName
++ change type to long in NutsExecCommand::getSleepMillis()/setSleepMillis(int sleepMillis);
 
 ## nuts 0.8.3.1 (PUBLISHED VERSION)
 - ```2022/02/01 	nuts 0.8.3.1``` released nuts-runtime-0.8.3.1.jar